+++ test/perspective-viewport.spec.js
@@ -1,22 +1,39 @@
 import test from 'tape-catch';
-import {Viewport, PerspectiveViewport, OrthographicViewport} from '../src';
+import {vec2} from 'gl-matrix';
+import {PerspectiveViewport} from '../src';
 
-const TEST_DATA = {
-  viewport: {
-    view: mat4.create(),
-    perspective: mat4.create()
+/* eslint-disable */
+const TEST_VIEWPORTS = [
+  {
+    mapState: {
+      width: 793,
+      height: 775,
+      latitude: 37.751537058389985,
+      longitude: -122.42694203247012,
+      zoom: 11.5
+    }
   },
-  perspective: {
-    eye: [10, 10, 10],
-    width: 793,
-    height: 775,
+  {
+    mapState: {
+      width: 793,
+      height: 775,
+      latitude: 20.751537058389985,
+      longitude: 22.42694203247012,
+      zoom: 15.5
+    }
   },
-  ortho: {
-    eye: [10, 10, 10],
-    width: 793,
-    height: 775
+  {
+    mapState: {
+      width: 793,
+      height: 775,
+      latitude: 50.751537058389985,
+      longitude: 42.42694203247012,
+      zoom: 15.5,
+      bearing: -44.48928121059271,
+      pitch: 43.670797287818566
+    }
   }
-};
+];
 
 test('PerspectiveViewport#imports', t => {
   t.ok(PerspectiveViewport, 'PerspectiveViewport import ok');
@@ -24,7 +41,93 @@ test('PerspectiveViewport#imports', t => {
 });
 
 test('PerspectiveViewport#constructor', t => {
-  t.ok(new PerspectiveViewport(TEST_DATA.perspective) instanceof PerspectiveViewport,
-    'Created new PerspectiveViewport with test args');
+  t.ok(new PerspectiveViewport() instanceof PerspectiveViewport,
+    'Created new PerspectiveViewport with default args');
+  t.end();
+});
+
+test('PerspectiveViewport#constructor - 0 width/height', t => {
+  const viewport = new PerspectiveViewport({
+    ...TEST_VIEWPORTS.mapState,
+    width: 0,
+    height: 0
+  });
+  t.ok(viewport instanceof PerspectiveViewport,
+    'PerspectiveViewport constructed successfully with 0 width and height');
+  t.end();
+});
+
+test('PerspectiveViewport.projectFlat', t => {
+  for (const vc of TEST_VIEWPORTS) {
+    const viewport = new PerspectiveViewport(vc.mapState);
+    for (const tc of TEST_VIEWPORTS) {
+      const {mapState} = tc;
+      const lnglatIn = [tc.mapState.longitude, tc.mapState.latitude];
+      const xy = viewport.projectFlat(lnglatIn);
+      const lnglat = viewport.unprojectFlat(xy);
+      t.comment(`Comparing [${lnglatIn}] to [${lnglat}]`);
+      t.ok(vec2.equals(lnglatIn, lnglat));
+    }
+  }
+  t.end();
+});
+
+test('PerspectiveViewport.project#3D', t => {
+  for (const vc of TEST_VIEWPORTS) {
+    const viewport = new PerspectiveViewport(vc.mapState);
+    for (const tc of TEST_VIEWPORTS) {
+      const {mapState} = tc;
+      const lnglatIn = [tc.mapState.longitude, tc.mapState.latitude, 0];
+      const xyz = viewport.project(lnglatIn);
+      const lnglat = viewport.unproject(xyz);
+      t.comment(`Comparing [${lnglatIn}] to [${lnglat}]`);
+      t.ok(vec2.equals(lnglatIn, lnglat));
+    }
+  }
+  t.end();
+});
+
+// TODO - this is not working at the moment
+
+// test('PerspectiveViewport.project#2D', t => {
+//   // Cross check positions
+//   for (const vc of TEST_VIEWPORTS) {
+//     const viewport = new PerspectiveViewport(vc.mapState);
+//     for (const tc of TEST_VIEWPORTS) {
+//       const {mapState} = tc;
+//       const lnglatIn = [tc.mapState.longitude, tc.mapState.latitude];
+//       const xy = viewport.project(lnglatIn);
+//       const lnglat = viewport.unproject(xy);
+//       t.comment(`Comparing [${lnglatIn}] to [${lnglat}]`);
+//       t.ok(vec2.equals(lnglatIn, lnglat));
+//     }
+//   }
+//   t.end();
+// });
+
+test('PerspectiveViewport.getScales', t => {
+  for (const vc of TEST_VIEWPORTS) {
+    const viewport = new PerspectiveViewport(vc.mapState);
+    const distanceScales = viewport.getDistanceScales();
+    t.ok(Array.isArray(distanceScales.metersPerPixel), 'metersPerPixel defined');
+    t.ok(Array.isArray(distanceScales.pixelsPerMeter), 'pixelsPerMeter defined');
+    t.ok(Array.isArray(distanceScales.degreesPerPixel), 'degreesPerPixel defined');
+    t.ok(Array.isArray(distanceScales.pixelsPerDegree), 'pixelsPerDegree defined');
+  }
+  t.end();
+});
+
+test('PerspectiveViewport.meterDeltas', t => {
+  for (const vc of TEST_VIEWPORTS) {
+    const viewport = new PerspectiveViewport(vc.mapState);
+    for (const tc of TEST_VIEWPORTS) {
+      const {mapState} = tc;
+      const coordinate = [tc.mapState.longitude, tc.mapState.latitude, 0];
+      const deltaLngLat = viewport.metersToLngLatDelta(coordinate);
+      const deltaMeters = viewport.lngLatDeltaToMeters(deltaLngLat);
+      t.comment(`Comparing [${deltaMeters}] to [${coordinate}]`);
+      t.ok(vec2.equals(deltaMeters, coordinate));
+    }
+  }
   t.end();
 });
